refactor(analysis): extract nesting complexity module
